+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.3.0 (unreleased)
 
 - Changed dataset directory to use `UserCacheDir`
+- Changed `seed` from `int64` to `uint64`
 - Dropped support for Go < 1.26
 
 ## 0.2.0 (2025-05-20)